完善添加任务部分的流程

+++ ResourceMonitor/ResourceMonitor/Models/DandanplayApi/AutoDownloadRule.cs
@@ -40,7 +40,7 @@ namespace ResourceMonitor.Models.DandanplayApi
         public string typeName { get; set; }
         public bool chooseNewerIfDuplicate { get; set; }
         public int limitFileSize { get; set; }
-        public int version { get; set; }
+        public long version { get; set; }
 
         public override string ToString()
         {
